Calculate per device batch size for transformer
